Corrected search panel so only updates if there is a large change in map delta
